Work around leftover children from cmd.exe in Windows 10
In Windows 10, if cmd.exe is running in a non-console context, when
it is terminated, its children do not automatically get terminated.
This is in contrast to the console context, and all situations on
previous versions of Windows.

We now terminate all of the direct children of cmd.exe in both
PTY and non-PTY cases, emulating the expected behavior of cmd.exe.
This only applies to processes launched via shell-host: we check
that supposed child processes have been spawned within the time
frame of our actual cmd.exe process before terminating them.